+++ CHANGELOG.md
@@ -1,39 +1,40 @@
 # Table of contents
-This changelog was generated with `version` software tool
+
 * [Overview](#overview)
 * [Statistics](#statistics)
 * [Version history](#version-history)
 * [Changelog](#changelog)
 # Overview
+
 Remote address:
 * **https://github.com/ErnestSzczepaniak/version**
+
+Branches:
+* **main [current]** (**16** commits)
+* test (**15** commits)
+
+Current version:
+* **0.6.0**
 
 Contributors:
 * **Ernest Szczepaniak**
 
-Branches:
-* **main [current]** (**15** commits)
-* test (**15** commits)
-
-Current version:
-* **0.5.1**
-
 # Statistics
+
 Project timeframe:
 * **20.04.2023 - 26.04.2023** (7 days)
 
 Code frequency:
-* **15** commits (**2.14** / day)
-* **10** files changed (**1.43** / day)
-* **425** insertions (**60.71** / day)
-* **376** deletions (**53.71** / day)
+* **16** commits (**2.29** / day)
 
 Commit structure:
-* **fix** - 6 (40.0%)
-* **feat** - 5 (33.33%)
-* **style** - 4 (26.67%)
+* **feat** - 6 (37.5%)
+* **fix** - 6 (37.5%)
+* **style** - 4 (25.0%)
 
 # Version history
+
+* [**0.6.0**](#060)
 * [**0.5.1**](#051)
 * [**0.5.0-rev.2**](#050-rev2)
 * [**0.5.0-rev.1**](#050-rev1)
@@ -51,10 +52,28 @@ Commit structure:
 * [**0.1.0**](#010)
 
 # Changelog
+
+## **[0.6.0](https://github.com/ErnestSzczepaniak/version/tree/ff646ce)**
+* **[26.04.2023]** [[ff646ce](https://github.com/ErnestSzczepaniak/version/commit/ff646ce)] (feat) - split changelog to sections (**Ernest Szczepaniak** @ 15:09:03)
+
+   `2 files changed (36 insertions, 16 deletions):`
+
+   * `CHANGELOG.md`
+   * `changlelog.py`
 ## **[0.5.1](https://github.com/ErnestSzczepaniak/version/tree/89e51aa)**
 * **[26.04.2023]** [[89e51aa](https://github.com/ErnestSzczepaniak/version/commit/89e51aa)] (fix) - prepare to merge (**Ernest Szczepaniak** @ 14:36:47)
+
+   `2 files changed (25 insertions, 20 deletions):`
+
+   * `CHANGELOG.md`
+   * `github.py`
 ## **[0.5.0-rev.2](https://github.com/ErnestSzczepaniak/version/tree/5864c0f)**
 * **[26.04.2023]** [[5864c0f](https://github.com/ErnestSzczepaniak/version/commit/5864c0f)] (style) - change *-rc.X to *-rev.X (**Ernest Szczepaniak** @ 14:17:21)
+
+   `2 files changed (18 insertions, 8 deletions):`
+
+   * `CHANGELOG.md`
+   * `github.py`
 ## **[0.5.0-rev.1](https://github.com/ErnestSzczepaniak/version/tree/6e61149)**
 * **[26.04.2023]** [[6e61149](https://github.com/ErnestSzczepaniak/version/commit/6e61149)] (feat) - add release candidates numbering (**Ernest Szczepaniak** @ 14:12:43)
 
@@ -64,10 +83,33 @@ Commit structure:
    release with same features (but with different styling, performance etc ...).
    ```
 
+
+   `3 files changed (37 insertions, 15 deletions):`
+
+   * `CHANGELOG.md`
+   * `commit.py`
+   * `github.py`
 ## **[0.4.0-rev.3](https://github.com/ErnestSzczepaniak/version/tree/fd6bd44)**
 * **[26.04.2023]** [[fd6bd44](https://github.com/ErnestSzczepaniak/version/commit/fd6bd44)] (style) - add tabs in COMMANDS dictionary (**Ernest Szczepaniak** @ 14:00:17)
+
+   `3 files changed (26 insertions, 15 deletions):`
+
+   * `CHANGELOG.md`
+   * `changlelog.py`
+   * `github.py`
 ## **[0.4.0-rev.2](https://github.com/ErnestSzczepaniak/version/tree/dd41587)**
 * **[26.04.2023]** [[dd41587](https://github.com/ErnestSzczepaniak/version/commit/dd41587)] (style) - rework to FORMAT_TABLE and COMMANDS (**Ernest Szczepaniak** @ 13:50:05)
+
+   `8 files changed (97 insertions, 93 deletions):`
+
+   * `CHANGELOG.md`
+   * `branch.py`
+   * `changlelog.py`
+   * `commit.py`
+   * `difference.py`
+   * `find.py`
+   * `github.py`
+   * `show.py`
 ## **[0.4.0-rev.1](https://github.com/ErnestSzczepaniak/version/tree/0c23698)**
 * **[21.04.2023]** [[0c23698](https://github.com/ErnestSzczepaniak/version/commit/0c23698)] (feat) - add branches in changelog (**Ernest Szczepaniak** @ 13:11:09)
 
@@ -75,10 +117,32 @@ Commit structure:
    this will help to clarify the repository structure
    ```
 
+
+   `4 files changed (161 insertions, 120 deletions):`
+
+   * `CHANGELOG.md`
+   * `branch.py`
+   * `changlelog.py`
+   * `github.py`
 ## **[0.3.2](https://github.com/ErnestSzczepaniak/version/tree/36f14bf)**
 * **[21.04.2023]** [[36f14bf](https://github.com/ErnestSzczepaniak/version/commit/36f14bf)] (fix) - switch to another branch (**Ernest Szczepaniak** @ 12:51:02)
+
+   `5 files changed (194 insertions, 97 deletions):`
+
+   * `CHANGELOG.md`
+   * `changlelog.py`
+   * `commit.py`
+   * `find.py`
+   * `github.py`
 ## **[0.3.1](https://github.com/ErnestSzczepaniak/version/tree/310b5b9)**
 * **[20.04.2023]** [[310b5b9](https://github.com/ErnestSzczepaniak/version/commit/310b5b9)] (fix) - try commit without body (**Ernest Szczepaniak** @ 12:53:39)
+
+   `4 files changed (41 insertions, 31 deletions):`
+
+   * `CHANGELOG.md`
+   * `changlelog.py`
+   * `commit.py`
+   * `github.py`
 ## **[0.3.0-rev.2](https://github.com/ErnestSzczepaniak/version/tree/6abd722)**
 * **[20.04.2023]** [[6abd722](https://github.com/ErnestSzczepaniak/version/commit/6abd722)] (style) - add styling (**Ernest Szczepaniak** @ 11:55:29)
 
@@ -90,6 +154,10 @@ Commit structure:
    dd2dadasd
    ```
 
+
+   `1 files changed (16 insertions, 0 deletions):`
+
+   * `CHANGELOG.md`
 ## **[0.3.0-rev.1](https://github.com/ErnestSzczepaniak/version/tree/9dab4b1)**
 * **[20.04.2023]** [[9dab4b1](https://github.com/ErnestSzczepaniak/version/commit/9dab4b1)] (feat) - check commit spacing (**Ernest Szczepaniak** @ 11:54:45)
 
@@ -106,6 +174,13 @@ Commit structure:
    6
    ```
 
+
+   `4 files changed (41 insertions, 50 deletions):`
+
+   * `CHANGELOG.md`
+   * `changlelog.py`
+   * `github.py`
+   * `markdown.py`
 ## **[0.2.2](https://github.com/ErnestSzczepaniak/version/tree/b48fadf)**
 * **[20.04.2023]** [[b48fadf](https://github.com/ErnestSzczepaniak/version/commit/b48fadf)] (fix) - verify changelog creation (**Ernest Szczepaniak** @ 08:56:10)
 
@@ -117,6 +192,15 @@ Commit structure:
    Than text afterwards.
    ```
 
+
+   `6 files changed (89 insertions, 347 deletions):`
+
+   * `CHANGELOG.md`
+   * `changlelog.py`
+   * `commit.py`
+   * `find.py`
+   * `github.py`
+   * `markdown.py`
 ## **[0.2.1](https://github.com/ErnestSzczepaniak/version/tree/6205f3d)**
 * **[20.04.2023]** [[6205f3d](https://github.com/ErnestSzczepaniak/version/commit/6205f3d)] (fix) - newline at end of body (**Ernest Szczepaniak** @ 08:20:47)
 
@@ -126,6 +210,10 @@ Commit structure:
    asd
    ```
 
+
+   `1 files changed (1 insertions, 0 deletions):`
+
+   * `main.py`
 ## **[0.2.0](https://github.com/ErnestSzczepaniak/version/tree/9b1e0a5)**
 * **[20.04.2023]** [[9b1e0a5](https://github.com/ErnestSzczepaniak/version/commit/9b1e0a5)] (feat) - test 2 line body (**Ernest Szczepaniak** @ 08:13:21)
 
@@ -135,6 +223,11 @@ Commit structure:
    qweqwe
    ```
 
+
+   `2 files changed (3 insertions, 0 deletions):`
+
+   * `github.py`
+   * `main.py`
 ## **[0.1.1](https://github.com/ErnestSzczepaniak/version/tree/86a3303)**
 * **[20.04.2023]** [[86a3303](https://github.com/ErnestSzczepaniak/version/commit/86a3303)] (fix) - remove empty spaces (**Ernest Szczepaniak** @ 08:03:50)
 
@@ -142,6 +235,11 @@ Commit structure:
    this will prevent from bad coding practices.
    ```
 
+
+   `2 files changed (1 insertions, 2 deletions):`
+
+   * `github.py`
+   * `main.py`
 ## **[0.1.0](https://github.com/ErnestSzczepaniak/version/tree/97beafd)**
 * **[20.04.2023]** [[97beafd](https://github.com/ErnestSzczepaniak/version/commit/97beafd)] (feat) - add hyperlinks to changelog (**Ernest Szczepaniak** @ 07:53:28)
 
@@ -149,3 +247,17 @@ Commit structure:
    This function works only when internet connection is present and github repository is a public one.
    ```
 
+
+   `11 files changed (690 insertions, 0 deletions):`
+
+   * `.gitignore`
+   * `CHANGELOG.md`
+   * `Makefile`
+   * `README.md`
+   * `changlelog.py`
+   * `commit.py`
+   * `find.py`
+   * `github.py`
+   * `main.py`
+   * `markdown.py`
+   * `show.py`